+++ Source/ACE.Entity/AceCharacter.cs
@@ -8,7 +8,7 @@ using System.Linq;
 namespace ACE.Entity
 {
     [DbTable("vw_ace_character")]
-    public class AceCharacter : AceObject
+    public class AceCharacter : AceObject, ICreatureStats
     {
         private readonly List<Friend> friends;
 
@@ -20,25 +20,25 @@ namespace ACE.Entity
 
             // Required default properties for character login
             // FIXME(ddevec): Should we have constants for (some of) these things?
-            ItemType = (uint)Enum.ItemType.Creature;
             AceObjectDescriptionFlags = (uint)(ObjectDescriptionFlag.Stuck | ObjectDescriptionFlag.Player | ObjectDescriptionFlag.Attackable);
-            WeenieHeaderFlags = (uint)(WeenieHeaderFlag.ItemCapacity | WeenieHeaderFlag.ContainerCapacity | WeenieHeaderFlag.Usable | WeenieHeaderFlag.RadarBehavior);
-            ItemsCapacity = 102;
-            ContainersCapacity = 7;
+            WeenieHeaderFlags = (uint)(WeenieHeaderFlag.ItemsCapacity | WeenieHeaderFlag.ContainersCapacity | WeenieHeaderFlag.Usable | WeenieHeaderFlag.RadarBehavior);
             WeenieClassId = 1;
-            Deleted = false;
-            DeleteTime = 0;
-            Level = 1;
-            AvailableExperience = 0;
-            TotalExperience = 0;
-            IconId = 100667446;
-            WeenieClassId = 1;
-            Burden = 0;
-            // SpellId = 0;
+            SetDataIdProperty(PropertyDataId.Icon, 100667446);
+            SetIntProperty(PropertyInt.ItemType, (uint)Enum.ItemType.Creature);
+            SetIntProperty(PropertyInt.ItemsCapacity, 102);
+            SetIntProperty(PropertyInt.ContainersCapacity, 7);
+            SetBoolProperty(PropertyBool.IsDeleted, false);
+            SetIntProperty(PropertyInt.TotalLogins, 0);
+            SetInt64Property(PropertyInt64.DeleteTime, 0);
+            SetIntProperty(PropertyInt.Level, 1);
+            SetInt64Property(PropertyInt64.AvailableExperience, 0);
+            SetInt64Property(PropertyInt64.TotalExperience, 0);
 
-            Radar = (byte)RadarBehavior.ShowAlways;
-            BlipColor = (byte)RadarColor.White;
-            ItemUseable = (uint)Usable.No;
+            SetIntProperty(PropertyInt.EncumbranceVal, 0);
+
+            SetIntProperty(PropertyInt.ShowableOnRadar, (byte)RadarBehavior.ShowAlways);
+            SetIntProperty(PropertyInt.RadarBlipColor, (byte)RadarColor.White);
+            SetIntProperty(PropertyInt.ItemUseable, (uint)Usable.No);
 
             SetBoolProperty(PropertyBool.FirstEnterWorldDone, false);
 
@@ -56,9 +56,9 @@ namespace ACE.Entity
             SetIntProperty(PropertyInt.NumDeaths, 0);
 
             SetIntProperty(PropertyInt.ChessRank, 1400);
-            SetIntProperty(PropertyInt.ChessTotalgames, 0);
-            SetIntProperty(PropertyInt.ChessGameslost, 0);
-            SetIntProperty(PropertyInt.ChessGameswon, 0);
+            SetIntProperty(PropertyInt.ChessTotalGames, 0);
+            SetIntProperty(PropertyInt.ChessGamesLost, 0);
+            SetIntProperty(PropertyInt.ChessGamesWon, 0);
             SetIntProperty(PropertyInt.FakeFishingSkill, 0);
 
             SetIntProperty(PropertyInt.PhysicsState, (uint)(Enum.PhysicsState.IgnoreCollision | Enum.PhysicsState.Gravity | Enum.PhysicsState.Hidden | Enum.PhysicsState.EdgeSlide));
@@ -135,7 +135,7 @@ namespace ACE.Entity
             set { SetIntProperty(PropertyInt.CharacterOptions2, value); }
         }
 
-        public uint TotalLogins
+        public new uint TotalLogins
         {
             get { return GetIntProperty(PropertyInt.TotalLogins) ?? 0; }
             set { SetIntProperty(PropertyInt.TotalLogins, value); }
@@ -143,7 +143,7 @@ namespace ACE.Entity
 
         public uint AccountId
         {
-            get { return InstanceIdProperties.Find(x => x.PropertyId == (uint)PropertyInstanceId.Account).PropertyValue.Value; }
+            get { return (uint)GetInstanceIdProperty(PropertyInstanceId.Account); }
             set { SetInstanceIdProperty(PropertyInstanceId.Account, value); }
         }
 
@@ -153,7 +153,7 @@ namespace ACE.Entity
             set { SetBoolProperty(PropertyBool.IsDeleted, value); }
         }
 
-        public ulong DeleteTime
+        public new ulong DeleteTime
         {
             get { return GetInt64Property(PropertyInt64.DeleteTime) ?? 0; }
             set { SetInt64Property(PropertyInt64.DeleteTime, value); }
@@ -197,106 +197,82 @@ namespace ACE.Entity
             dirtyOptions = true;
         }
 
-        public ulong AvailableExperience
+        public new ulong AvailableExperience
         {
             get { return GetInt64Property(PropertyInt64.AvailableExperience) ?? 0; }
             set { SetInt64Property(PropertyInt64.AvailableExperience, value); }
         }
 
-        public ulong TotalExperience
+        public new ulong TotalExperience
         {
             get { return GetInt64Property(PropertyInt64.TotalExperience) ?? 0; }
             set { SetInt64Property(PropertyInt64.TotalExperience, value); }
         }
 
-        public uint Age
+        public new uint Age
         {
             get { return GetIntProperty(PropertyInt.Age) ?? 0; }
             set { SetIntProperty(PropertyInt.Age, value); }
         }
 
-        public bool IsDeleted
+        public new bool IsDeleted
         {
             get { return GetBoolProperty(PropertyBool.IsDeleted) ?? false; }
             set { SetBoolProperty(PropertyBool.IsDeleted, value); }
         }
 
-        public ulong DeletedTime
-        {
-            get { return GetInt64Property(PropertyInt64.DeleteTime) ?? 0; }
-            set { SetInt64Property(PropertyInt64.DeleteTime, value); }
-        }
-
-        public string DateOfBirth
-        {
-            get { return GetStringProperty(PropertyString.DateOfBirth); }
-            set { SetStringProperty(PropertyString.DateOfBirth, value); }
-        }
-
-        public uint AvailableSkillCredits
+        public new uint AvailableSkillCredits
         {
             get { return GetIntProperty(PropertyInt.AvailableSkillCredits) ?? 0; }
             set { SetIntProperty(PropertyInt.AvailableSkillCredits, value); }
         }
 
-        public uint TotalSkillCredits
+        public new uint TotalSkillCredits
         {
             get { return GetIntProperty(PropertyInt.TotalSkillCredits) ?? 0; }
             set { SetIntProperty(PropertyInt.TotalSkillCredits, value); }
         }
 
-        public uint NumDeaths
+        public new uint NumDeaths
         {
             get { return GetIntProperty(PropertyInt.NumDeaths) ?? 0; }
             set { SetIntProperty(PropertyInt.NumDeaths, value); }
         }
 
-        public uint DeathLevel
+        public new uint DeathLevel
         {
             get { return GetIntProperty(PropertyInt.DeathLevel) ?? 0; }
             set { SetIntProperty(PropertyInt.DeathLevel, value); }
         }
 
-        public uint VitaeCpPool
+        public new uint VitaeCpPool
         {
             get { return GetIntProperty(PropertyInt.VitaeCpPool) ?? 0; }
             set { SetIntProperty(PropertyInt.VitaeCpPool, value); }
         }
 
-        public bool IsAdmin
+        public new bool IsAdmin
         {
             get { return GetBoolProperty(PropertyBool.IsAdmin) ?? false; }
             set { SetBoolProperty(PropertyBool.IsAdmin, value); }
         }
 
-        public bool IsEnvoy
+        public new bool IsEnvoy
         {
             get { return GetBoolProperty(PropertyBool.IsSentinel) ?? false; }
             set { SetBoolProperty(PropertyBool.IsSentinel, value); }
         }
 
-        public bool IsArch
+        public new bool IsArch
         {
             get { return GetBoolProperty(PropertyBool.IsArch) ?? false; }
             set { SetBoolProperty(PropertyBool.IsArch, value); }
         }
 
-        public bool IsPsr
+        public new bool IsPsr
         {
             get { return GetBoolProperty(PropertyBool.IsPsr) ?? false; }
             set { SetBoolProperty(PropertyBool.IsPsr, value); }
-        }
-
-        public uint Heritage
-        {
-            get { return GetIntProperty(PropertyInt.HeritageGroup).Value; }
-            set { SetIntProperty(PropertyInt.HeritageGroup, value); }
-        }
-
-        public uint Gender
-        {
-            get { return GetIntProperty(PropertyInt.Gender).Value; }
-            set { SetIntProperty(PropertyInt.Gender, value); }
         }
 
         public uint EyesTexture
@@ -371,6 +347,54 @@ namespace ACE.Entity
             set { SetDataIdProperty(PropertyDataId.EyesPalette, value); }
         }
 
+        public uint? SetupTableId
+        {
+            get { return GetDataIdProperty(PropertyDataId.Setup); }
+            set { SetDataIdProperty(PropertyDataId.Setup, value); }
+        }
+
+        public uint? MotionTableId
+        {
+            get { return GetDataIdProperty(PropertyDataId.MotionTable); }
+            set { SetDataIdProperty(PropertyDataId.MotionTable, value); }
+        }
+
+        public ushort? PhysicsScript
+        {
+            get { return (ushort?)GetDataIdProperty(PropertyDataId.PhysicsScript); }
+            set { SetDataIdProperty(PropertyDataId.PhysicsScript, value); }
+        }
+
+        public uint? PhysicsTableId
+        {
+            get { return GetDataIdProperty(PropertyDataId.PhysicsEffectTable); }
+            set { SetDataIdProperty(PropertyDataId.PhysicsEffectTable, value); }
+        }
+
+        public uint? SoundTableId
+        {
+            get { return GetDataIdProperty(PropertyDataId.SoundTable); }
+            set { SetDataIdProperty(PropertyDataId.SoundTable, value); }
+        }
+
+        public uint? CombatTableId
+        {
+            get { return GetDataIdProperty(PropertyDataId.CombatTable); }
+            set { SetDataIdProperty(PropertyDataId.CombatTable, value); }
+        }
+
+        public new uint Level
+        {
+            get { return GetIntProperty(PropertyInt.Level) ?? 1; }
+            set { SetIntProperty(PropertyInt.Level, value); }
+        }
+
+        public uint? PaletteId
+        {
+            get { return GetDataIdProperty(PropertyDataId.PaletteBase); }
+            set { SetDataIdProperty(PropertyDataId.PaletteBase, value); }
+        }
+
         private bool dirtyOptions = true;
 
         private Dictionary<CharacterOption, bool> allOptions = new Dictionary<CharacterOption, bool>();
@@ -406,24 +430,6 @@ namespace ACE.Entity
 
                 return new ReadOnlyDictionary<CharacterOption, bool>(allOptions);
             }
-        }
-
-        public Position LastPortal
-        {
-            get { return GetPosition(PositionType.LastPortal); }
-            set { SetPosition(PositionType.LastPortal, value); }
-        }
-
-        public Position Sanctuary
-        {
-            get { return GetPosition(PositionType.Sanctuary); }
-            set { SetPosition(PositionType.Sanctuary, value); }
-        }
-
-        public Position LastOutsideDeath
-        {
-            get { return GetPosition(PositionType.LastOutsideDeath); }
-            set { SetPosition(PositionType.LastOutsideDeath, value); }
         }
 
         public void SetCharacterOption(CharacterOption option, bool value)