+++ Source/ACE/Entity/Generator.cs
@@ -1,4 +1,5 @@
 ﻿using ACE.Entity.Enum;
+using ACE.Entity.Enum.Properties;
 using ACE.Network.Enum;
 using ACE.Network.Motion;
 using System;
@@ -31,5 +32,95 @@ namespace ACE.Entity
         ////    WeenieClassId = aceO.WeenieClassId;
         ////    GameDataType = aceO.Type;
         ////}
+
+        public bool GeneratorStatus
+        {
+            get { return AceObject.GeneratorStatus ?? false; }
+            set { AceObject.GeneratorStatus = value; }
+        }
+
+        public bool GeneratorEnteredWorld
+        {
+            get { return AceObject.GeneratorEnteredWorld ?? false; }
+            set { AceObject.GeneratorEnteredWorld = value; }
+        }
+
+        public bool GeneratorDisabled
+        {
+            get { return AceObject.GeneratorDisabled ?? false; }
+            set { AceObject.GeneratorDisabled = value; }
+        }
+
+        public bool GeneratedTreasureItem
+        {
+            get { return AceObject.GeneratedTreasureItem ?? false; }
+            set { AceObject.GeneratedTreasureItem = value; }
+        }
+
+        public bool GeneratorAutomaticDestruction
+        {
+            get { return AceObject.GeneratorAutomaticDestruction ?? false; }
+            set { AceObject.GeneratorAutomaticDestruction = value; }
+        }
+
+        public bool CanGenerateRare
+        {
+            get { return AceObject.CanGenerateRare ?? false; }
+            set { AceObject.CanGenerateRare = value; }
+        }
+
+        public bool CorpseGeneratedRare
+        {
+            get { return AceObject.CorpseGeneratedRare ?? false; }
+            set { AceObject.CorpseGeneratedRare = value; }
+        }
+
+        public bool SuppressGenerateEffect
+        {
+            get { return AceObject.SuppressGenerateEffect ?? false; }
+            set { AceObject.SuppressGenerateEffect = value; }
+        }
+
+        public bool ChestRegenOnClose
+        {
+            get { return AceObject.ChestRegenOnClose ?? false; }
+            set { AceObject.ChestRegenOnClose = value; }
+        }
+
+        public bool ChestClearedWhenClosed
+        {
+            get { return AceObject.ChestClearedWhenClosed ?? false; }
+            set { AceObject.ChestClearedWhenClosed = value; }
+        }
+
+        public uint GeneratorTimeType
+        {
+            get { return AceObject.GeneratorTimeType ?? 0; }
+            set { AceObject.GeneratorTimeType = value; }
+        }
+
+        public uint GeneratorProbability
+        {
+            get { return AceObject.GeneratorProbability ?? 0; }
+            set { AceObject.GeneratorProbability = value; }
+        }
+
+        public uint MaxGeneratedObjects
+        {
+            get { return AceObject.MaxGeneratedObjects ?? 0; }
+            set { AceObject.MaxGeneratedObjects = value; }
+        }
+
+        public uint GeneratorType
+        {
+            get { return AceObject.GeneratorType ?? 0; }
+            set { AceObject.GeneratorType = value; }
+        }
+
+        public uint ActivationCreateClass
+        {
+            get { return AceObject.ActivationCreateClass ?? 0; }
+            set { AceObject.ActivationCreateClass = value; }
+        }
     }
 }