+++ Source/ACE/Factories/CorpseObjectFactory.cs
@@ -16,10 +16,10 @@ namespace ACE.Factories
         public static Container CreateCorpse(Creature template, Position newPosition)
         {
             ushort wcidCorpse = 21;
-            WeenieHeaderFlag weenie = WeenieHeaderFlag.ItemCapacity | WeenieHeaderFlag.ContainerCapacity | WeenieHeaderFlag.Usable | WeenieHeaderFlag.UseRadius | WeenieHeaderFlag.Burden;
+            WeenieHeaderFlag weenie = WeenieHeaderFlag.ItemsCapacity | WeenieHeaderFlag.ContainersCapacity | WeenieHeaderFlag.Usable | WeenieHeaderFlag.UseRadius | WeenieHeaderFlag.Burden;
             ObjectDescriptionFlag objDesc = ObjectDescriptionFlag.Openable | ObjectDescriptionFlag.Stuck | ObjectDescriptionFlag.Attackable | ObjectDescriptionFlag.Corpse; // = bitfield 8213
             string name = $"Corpse of {template.Name}";
-            Container wo = new Container(template.AceCorpse, new ObjectGuid(CommonObjectFactory.DynamicObjectId, GuidType.None), name, wcidCorpse, objDesc, weenie, newPosition);
+            Container wo = new Container(template.AceCorpse, GuidManager.NewItemGuid(), name, wcidCorpse, objDesc, weenie, newPosition);
 
             // TODO: Find the correct motionstate to create a corpse with. For now only the dead motionstate works
             // wo.CurrentMotionState = new GeneralMotion(MotionStance.Standing);
@@ -27,11 +27,11 @@ namespace ACE.Factories
             wo.SoundTableId = 536871106; // SoundTableId in DB - constant value according to pcap
             wo.PhysicsTableId = 872415342; // phstableid in DB - constant value according to pcap
 
-            wo.PhysicsDescriptionFlag = PhysicsDescriptionFlag.CSetup | PhysicsDescriptionFlag.MTable | PhysicsDescriptionFlag.ObjScale | PhysicsDescriptionFlag.Stable | PhysicsDescriptionFlag.Petable | PhysicsDescriptionFlag.Position | PhysicsDescriptionFlag.Movement; // 104579 - according to pcap
+            wo.PhysicsDescriptionFlag = PhysicsDescriptionFlag.CSetup | PhysicsDescriptionFlag.MTable | PhysicsDescriptionFlag.ObjScale | PhysicsDescriptionFlag.STable | PhysicsDescriptionFlag.PeTable | PhysicsDescriptionFlag.Position | PhysicsDescriptionFlag.Movement; // 104579 - according to pcap
             wo.PhysicsState = PhysicsState.Ethereal | PhysicsState.IgnoreCollision | PhysicsState.Gravity; // 1044 - according to pcap
 
             uint tmpIcon = 100667504;
-            wo.Icon = tmpIcon;
+            wo.IconId = tmpIcon;
 
             wo.ItemCapacity = 120; // constant value according to pcap
             wo.ContainerCapacity = 10; // constant value according to pcap